server/requirements.txt
Django==1.8.2
django-localflavor==1.1
djangorestframework==3.1.3
djangorestframework-gis==0.9.1
psycopg2==2.6.1
wheel==0.24.0
